requirements.txt
pandas==2.2.3
pyarrow==18.1.0
numpy==2.1.3
SQLAlchemy
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.4
ruff==0.8.4
dbt-postgres==1.10.0